Little tidy up after rebase
